+++ ultimate/testData/inspections/spring/componentScan/foo/beans.kt
@@ -1,0 +1,6 @@
+package foo
+
+// Make PsiPackage not empty
+fun dummyFun() {
+
+}